added train_test function

+++ preprocessing.py
@@ -82,22 +82,22 @@ def load_data_discrete():
     x = list(reader)
     dataset = np.array(x[1:])
     dataset = pd.DataFrame(data = {
-                        'acceleration_mean' : [float(dataset[i][0]) for i in range(len(dataset))],
-                        'acceleration_stdev' : [float(dataset[i][1]) for i in range(len(dataset))],
-                        'pitch1' : [float(dataset[i][2]) for i in range(len(dataset))],
-                        'pitch2' : [float(dataset[i][3]) for i in range(len(dataset))],
-                        'pitch3' : [float(dataset[i][4]) for i in range(len(dataset))],
-                        'roll1' : [float(dataset[i][5]) for i in range(len(dataset))],
-                        'roll2' : [float(dataset[i][6]) for i in range(len(dataset))],
-                        'roll3' : [float(dataset[i][7]) for i in range(len(dataset))],
+                        'acceleration_mean' : [int(dataset[i][0]) for i in range(len(dataset))],
+                        'acceleration_stdev' : [int(dataset[i][1]) for i in range(len(dataset))],
+                        'pitch1' : [int(dataset[i][2]) for i in range(len(dataset))],
+                        'pitch2' : [int(dataset[i][3]) for i in range(len(dataset))],
+                        'pitch3' : [int(dataset[i][4]) for i in range(len(dataset))],
+                        'roll1' : [int(dataset[i][5]) for i in range(len(dataset))],
+                        'roll2' : [int(dataset[i][6]) for i in range(len(dataset))],
+                        'roll3' : [int(dataset[i][7]) for i in range(len(dataset))],
                         'sitting' : [int(dataset[i][8]) for i in range(len(dataset))],
                         'sittingdown' : [int(dataset[i][9]) for i in range(len(dataset))],
                         'standing' : [int(dataset[i][10]) for i in range(len(dataset))],
                         'standingup' : [int(dataset[i][11]) for i in range(len(dataset))],
                         'walking' : [int(dataset[i][12]) for i in range(len(dataset))],
-                        'total_accel_sensor_1' : [float(dataset[i][13]) for i in range(len(dataset))],
-                        'total_accel_sensor_2' : [float(dataset[i][14]) for i in range(len(dataset))],
-                        'total_accel_sensor_4' : [float(dataset[i][15]) for i in range(len(dataset))]})
+                        'total_accel_sensor_1' : [int(dataset[i][13]) for i in range(len(dataset))],
+                        'total_accel_sensor_2' : [int(dataset[i][14]) for i in range(len(dataset))],
+                        'total_accel_sensor_4' : [int(dataset[i][15]) for i in range(len(dataset))]})
 
     return dataset
